requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
bcrypt==4.0.1
gunicorn==21.2.0
